(sql)-ln support
Aligned the use of log() function to sql definition of ln
