+++ core/src/main/java/io/questdb/griffin/engine/functions/math/LnDoubleFunctionFactory.java
@@ -34,10 +34,10 @@ import io.questdb.griffin.engine.functions.UnaryFunction;
 import io.questdb.std.IntList;
 import io.questdb.std.ObjList;
 
-public class LogDoubleFunctionFactory implements FunctionFactory {
+public class LnDoubleFunctionFactory implements FunctionFactory {
     @Override
     public String getSignature() {
-        return "log(D)";
+        return "ln(D)";
     }
 
     @Override
@@ -46,13 +46,13 @@ public class LogDoubleFunctionFactory implements FunctionFactory {
             IntList argPositions, CairoConfiguration configuration,
             SqlExecutionContext sqlExecutionContext
     ) {
-        return new LogFunction(args.getQuick(0));
+        return new LnFunction(args.getQuick(0));
     }
 
-    private static class LogFunction extends DoubleFunction implements UnaryFunction {
+    private static class LnFunction extends DoubleFunction implements UnaryFunction {
         final Function function;
 
-        public LogFunction(Function function) {
+        public LnFunction(Function function) {
             this.function = function;
         }
